requirements.txt
django==2.1
django-bootstrap4
coverage>=4.4
django_coverage_plugin
psycopg2-binary
gunicorn
django-crispy-forms
django-import-export
django-autocomplete-light
parsel
